Refactor et navigation à l'ajout
