+++ src/components/view/localCalendar.tsx
@@ -1,11 +1,22 @@
-import React, { useContext } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import { Calendar, momentLocalizer } from 'react-big-calendar';
 import moment from 'moment';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
-import { CalModelContext } from '@/model/calModel';
+import { CalModelContext } from '@/components/model/calModel';
+import {findEarliestEventDate} from '../util/calEventOperations'
 
 const LocalCalendar: React.FC = () => {
     const {events} = useContext(CalModelContext);
+    const [selectedDate, setSelectedDate] = useState<Date>(findEarliestEventDate(events));
+
+    
+    useEffect(() => {
+        setSelectedDate(findEarliestEventDate(events));
+    }, [events]);
+
+    const onNavigate = (newDate: Date) => {
+        setSelectedDate(newDate);
+    }
 
     const localizer = momentLocalizer(moment);
 
@@ -17,6 +28,8 @@ const LocalCalendar: React.FC = () => {
                 startAccessor="start"
                 endAccessor="end"
                 style={{ height: 500 }}
+                date={selectedDate}
+                onNavigate={onNavigate}
             />
         </div>
     );
